Binomial Test

Test whether the proportion of successes on a two-level categorical dependent variable significantly differs from a hypothesized value.

infer_binom_calc(n, success, prob = 0.5, ...)

infer_binom_test(data, variable, prob = 0.5)

Arguments

n
number of observations

success
number of successes

prob
assumed probability of success on a trial

...
additional arguments passed to or from other methods

data
a data.frame or a tibble

variable
factor; column in data

Value

binom_test returns an object of class "binom_test". An object of class "binom_test" is a list containing the following components:

n

number of observations

k

number of successes

exp_k

expected number of successes

obs_p

assumed probability of success

exp_p

expected probability of success

lower

lower one sided p value

upper

upper one sided p value

Deprecated Functions

binom_calc() and binom_test() have been deprecated. Instead use infer_binom_cal() and infer_binom_test().

References

Hoel, P. G. 1984. Introduction to Mathematical Statistics. 5th ed. New York: Wiley.

See also

binom.test

Examples

# using calculator
infer_binom_calc(32, 13, prob = 0.5)

#>             Binomial Test              
#>  --------------------------------------
#>   Group    N     Obs. Prop    Exp. Prop 
#>  --------------------------------------
#>     0      19      0.59375        0.500 
#>     1      13      0.40625        0.500 
#>  --------------------------------------
#> 
#> 
#>                  Test Summary                 
#>  --------------------------------------------
#>   Tail              Prob             p-value  
#>  --------------------------------------------
#>   Lower         Pr(k <= 13)          0.188543 
#>   Upper         Pr(k >= 13)          0.892336 
#>  --------------------------------------------

# using data set
infer_binom_test(hsb, female, prob = 0.5)

#>              Binomial Test              
#>  ---------------------------------------
#>   Group     N     Obs. Prop    Exp. Prop 
#>  ---------------------------------------
#>     0       91        0.455        0.500 
#>     1      109        0.545        0.500 
#>  ---------------------------------------
#> 
#> 
#>                   Test Summary                  
#>  ----------------------------------------------
#>   Tail               Prob              p-value  
#>  ----------------------------------------------
#>   Lower          Pr(k <= 109)          0.910518 
#>   Upper          Pr(k >= 109)          0.114623 
#>  ----------------------------------------------