Two Sample Variance Comparison Test

var_test performs tests on the equality of standard deviations (variances).

var_test(variable, ..., group_var = NA, alternative = c("less", "greater",
  "all"))

Arguments

variable

a numeric vector

...

numeric vectors

group_var

a grouping variable

alternative

a character string specifying the alternative hypothesis, must be one of "both" (default), "greater", "less" or "all". You can specify just the initial letter.

Value

var_test returns an object of class "var_test". An object of class "var_test" is a list containing the following components:

References

Sheskin, D. J. 2007. Handbook of Parametric and Nonparametric Statistical Procedures, 4th edition. : Chapman & Hall/CRC.

See also

var.test

Examples

# using grouping variable
# lower tail
var_test(mtcars$mpg, group_var = mtcars$vs, alternative = 'less')

#>                Variance Ratio Test                 
#> --------------------------------------------------
#>   Group      Obs    Mean     Std. Err.    Std. Dev. 
#> --------------------------------------------------
#>    0        18     16.62      0.91         3.86    
#>    1        14     24.56      1.44         5.38    
#> --------------------------------------------------
#>  combined    32     20.09      1.07         6.03    
#> --------------------------------------------------
#> 
#>            Lower Tail Test           
#>            ---------------           
#>          ratio = sd(0) / (1)         
#>             Ho: ratio = 1            
#>             Ha: ratio < 1             
#> 
#>         Variance Ratio Test          
#> ------------------------------------
#>    F       Num DF    Den DF      p    
#> ------------------------------------
#>  0.5151      17        13      0.0999 
#> ------------------------------------

# upper tail
var_test(mtcars$mpg, group_var = mtcars$vs, alternative = 'greater')

#>                Variance Ratio Test                 
#> --------------------------------------------------
#>   Group      Obs    Mean     Std. Err.    Std. Dev. 
#> --------------------------------------------------
#>    0        18     16.62      0.91         3.86    
#>    1        14     24.56      1.44         5.38    
#> --------------------------------------------------
#>  combined    32     20.09      1.07         6.03    
#> --------------------------------------------------
#> 
#>            Upper Tail Test           
#>            ---------------           
#>             Ho: ratio = 1            
#>             Ha: ratio > 1             
#> 
#>         Variance Ratio Test          
#> ------------------------------------
#>    F       Num DF    Den DF      p    
#> ------------------------------------
#>  0.5151      17        13      0.9001 
#> ------------------------------------

# all tails
var_test(mtcars$mpg, group_var = mtcars$vs, alternative = 'all')

#>                Variance Ratio Test                 
#> --------------------------------------------------
#>   Group      Obs    Mean     Std. Err.    Std. Dev. 
#> --------------------------------------------------
#>    0        18     16.62      0.91         3.86    
#>    1        14     24.56      1.44         5.38    
#> --------------------------------------------------
#>  combined    32     20.09      1.07         6.03    
#> --------------------------------------------------
#> 
#>                 Variance Ratio Test                 
#> --------------------------------------------------
#>         F              Num DF           Den DF      
#> --------------------------------------------------
#>       0.5151             17               13        
#> --------------------------------------------------
#> 
#>        Null & Alternate Hypothesis        
#> ----------------------------------------
#>           ratio = sd(0) / (1)           
#>               Ho: ratio = 1              
#> 
#>     Ha: ratio < 1        Ha: ratio > 1    
#>   Pr(F < f) = 0.0999   Pr(F > f) = 0.9001 
#> ----------------------------------------

# using two variables
# lower tail
var_test(hsb$read, hsb$write, alternative = 'less')

#>                Variance Ratio Test                 
#> --------------------------------------------------
#>   Group      Obs    Mean     Std. Err.    Std. Dev. 
#> --------------------------------------------------
#>   read      200    52.23      0.72         10.25   
#>  write      200    52.77      0.67         9.48    
#> --------------------------------------------------
#>  combined    400    52.5       0.49         9.86    
#> --------------------------------------------------
#> 
#>            Lower Tail Test           
#>            ---------------           
#>       ratio = sd(read) / (write)     
#>             Ho: ratio = 1            
#>             Ha: ratio < 1             
#> 
#>         Variance Ratio Test          
#> ------------------------------------
#>    F       Num DF    Den DF      p    
#> ------------------------------------
#>  1.1701     199       199      0.8656 
#> ------------------------------------

# upper tail
var_test(hsb$read, hsb$write, alternative = 'greater')

#>                Variance Ratio Test                 
#> --------------------------------------------------
#>   Group      Obs    Mean     Std. Err.    Std. Dev. 
#> --------------------------------------------------
#>   read      200    52.23      0.72         10.25   
#>  write      200    52.77      0.67         9.48    
#> --------------------------------------------------
#>  combined    400    52.5       0.49         9.86    
#> --------------------------------------------------
#> 
#>            Upper Tail Test           
#>            ---------------           
#>             Ho: ratio = 1            
#>             Ha: ratio > 1             
#> 
#>         Variance Ratio Test          
#> ------------------------------------
#>    F       Num DF    Den DF      p    
#> ------------------------------------
#>  1.1701     199       199      0.1344 
#> ------------------------------------

# all tails
var_test(hsb$read, hsb$write, alternative = 'all')

#>                Variance Ratio Test                 
#> --------------------------------------------------
#>   Group      Obs    Mean     Std. Err.    Std. Dev. 
#> --------------------------------------------------
#>   read      200    52.23      0.72         10.25   
#>  write      200    52.77      0.67         9.48    
#> --------------------------------------------------
#>  combined    400    52.5       0.49         9.86    
#> --------------------------------------------------
#> 
#>                 Variance Ratio Test                 
#> --------------------------------------------------
#>         F              Num DF           Den DF      
#> --------------------------------------------------
#>       1.1701            199              199        
#> --------------------------------------------------
#> 
#>        Null & Alternate Hypothesis        
#> ----------------------------------------
#>        ratio = sd(read) / (write)       
#>               Ho: ratio = 1              
#> 
#>     Ha: ratio < 1        Ha: ratio > 1    
#>   Pr(F < f) = 0.8656   Pr(F > f) = 0.1344 
#> ----------------------------------------