Paired t test

infer_ts_paired_ttest tests that two samples have the same mean, assuming paired data.

infer_ts_paired_ttest(data, x, y, confint = 0.95, alternative = c("both",
  "less", "greater", "all"))

Arguments

data
a data.frame or tibble

x
numeric; column in data

y
numeric; column in data

confint
confidence level

alternative
a character string specifying the alternative hypothesis, must be one of "both" (default), "greater", "less" or "all". You can specify just the initial letter.

Value

infer_ts_paired_ttest returns an object of class "infer_ts_paired_ttest". An object of class "infer_ts_paired_ttest" is a list containing the following components:

Obs

number of observations

b

mean, standard deviation and standard error of x, y and their difference

tstat

t statistic

p_lower

lower one-sided p-value

p_upper

upper one-sided p-value

p_two_tail

two sided p-value

corr

Correlation of x and y

corsig

p-value of correlation test

conf_int1

confidence interval for mean of x

conf_int2

confidence interval for mean of y

conf_int_diff

confidence interval for mean of difference of x and y

df

degrees of freedom

confint

confidence level

alternative

alternative hypothesis

var_names

names of x and y

xy

string used in printing results of the test

Deprecated Function

paired_ttest() has been deprecated. Instead use infer_ts_paired_ttest().

References

Sheskin, D. J. 2007. Handbook of Parametric and Nonparametric Statistical Procedures, 4th edition. : Chapman & Hall/CRC.

See also

t.test

Examples

# lower tail
infer_ts_paired_ttest(hsb, read, write, alternative = 'less')

#>                          Paired Samples Statistics                          
#> ---------------------------------------------------------------------------
#> Variables    Obs    Mean     Std. Err.    Std. Dev.    [95% Conf. Interval] 
#> ---------------------------------------------------------------------------
#>    read       200    52.23      0.72         10.25         50.8      53.66   
#>    write      200    52.77      0.67         9.48         51.45      54.09    
#> ---------------------------------------------------------------------------
#>    diff       200    -0.55      0.63         8.89         -1.79       0.69    
#> ---------------------------------------------------------------------------
#> 
#>          Paired Samples Correlations         
#> -------------------------------------------
#>   Variables      Obs    Correlation    Sig.
#>  read & write    200       0.60        0 
#> -------------------------------------------
#> 
#>           Paired Samples Test           
#>           -------------------           
#>       Ho: mean(read - write) = 0        
#>       Ha: mean(read - write) < 0        
#> 
#> ---------------------------------------
#>   Variables        t       df     Sig.  
#> ---------------------------------------
#>  read - write    -0.873    199    0.192 
#> ---------------------------------------

# upper tail
infer_ts_paired_ttest(hsb, read, write, alternative = 'greater')

#>                          Paired Samples Statistics                          
#> ---------------------------------------------------------------------------
#> Variables    Obs    Mean     Std. Err.    Std. Dev.    [95% Conf. Interval] 
#> ---------------------------------------------------------------------------
#>    read       200    52.23      0.72         10.25         50.8      53.66   
#>    write      200    52.77      0.67         9.48         51.45      54.09    
#> ---------------------------------------------------------------------------
#>    diff       200    -0.55      0.63         8.89         -1.79       0.69    
#> ---------------------------------------------------------------------------
#> 
#>          Paired Samples Correlations         
#> -------------------------------------------
#>   Variables      Obs    Correlation    Sig.
#>  read & write    200       0.60        0 
#> -------------------------------------------
#> 
#>           Paired Samples Test           
#>           -------------------           
#>       Ho: mean(read - write) = 0        
#>       Ha: mean(read - write) > 0        
#> 
#> ---------------------------------------
#>   Variables        t       df     Sig.  
#> ---------------------------------------
#>  read - write    -0.873    199    0.808 
#> ---------------------------------------

# both tails
infer_ts_paired_ttest(hsb, read, write, alternative = 'both')

#>                          Paired Samples Statistics                          
#> ---------------------------------------------------------------------------
#> Variables    Obs    Mean     Std. Err.    Std. Dev.    [95% Conf. Interval] 
#> ---------------------------------------------------------------------------
#>    read       200    52.23      0.72         10.25         50.8      53.66   
#>    write      200    52.77      0.67         9.48         51.45      54.09    
#> ---------------------------------------------------------------------------
#>    diff       200    -0.55      0.63         8.89         -1.79       0.69    
#> ---------------------------------------------------------------------------
#> 
#>          Paired Samples Correlations         
#> -------------------------------------------
#>   Variables      Obs    Correlation    Sig.
#>  read & write    200       0.60        0 
#> -------------------------------------------
#> 
#>           Paired Samples Test           
#>           -------------------           
#>       Ho: mean(read - write) = 0        
#>       Ha: mean(read - write) ~= 0       
#> 
#> ---------------------------------------
#>   Variables        t       df     Sig.  
#> ---------------------------------------
#>  read - write    -0.873    199    0.384 
#> ---------------------------------------

# all tails
infer_ts_paired_ttest(hsb, read, write, alternative = 'all')

#>                          Paired Samples Statistics                          
#> ---------------------------------------------------------------------------
#> Variables    Obs    Mean     Std. Err.    Std. Dev.    [95% Conf. Interval] 
#> ---------------------------------------------------------------------------
#>    read       200    52.23      0.72         10.25         50.8      53.66   
#>    write      200    52.77      0.67         9.48         51.45      54.09    
#> ---------------------------------------------------------------------------
#>    diff       200    -0.55      0.63         8.89         -1.79       0.69    
#> ---------------------------------------------------------------------------
#> 
#>          Paired Samples Correlations         
#> -------------------------------------------
#>   Variables      Obs    Correlation    Sig.
#>  read & write    200       0.60        0 
#> -------------------------------------------
#> 
#>                 Ho: mean(read - write) = mean(diff) = 0                  
#> 
#>    Ha: mean(diff) < 0      Ha: mean(diff) ~= 0       Ha: mean(diff) > 0    
#>        t = -0.873               t = -0.873               t = -0.873        
#>      P < t = 0.192           P > |t| = 0.384           P > t = 0.808